Major upgrade to load times
Load times have been reduced with a combination of lazy loading, and swapping out modules for counterparts that support tree shaking. Also started work on the sorting system again, but are now trying to sort the json directly. This has failed. For unknow reasons the sorting method fails to catch all of the items that meet is criteria.